+++ lib/file-parsers/test/TestSuite/GeneratedTests.hs
@@ -5,7 +5,7 @@ module TestSuite.GeneratedTests
 import           Test.Tasty                     (TestTree, testGroup)
 import qualified TestSuite.GeneratedTests.Fasta as Fasta (testSuite)
 import qualified TestSuite.GeneratedTests.Fastc as Fastc (testSuite)
-import qualified TestSuite.GeneratedTests.Nexus as Nexus (testSuite)
+--import qualified TestSuite.GeneratedTests.Nexus as Nexus (testSuite)
 import qualified TestSuite.GeneratedTests.TNT   as TNT (testSuite)
 
 testSuite :: IO TestTree
@@ -13,6 +13,6 @@ testSuite = testGroup "Dynamically generated tests"
            <$> sequence
              [ Fasta.testSuite
              , Fastc.testSuite
-             , Nexus.testSuite
+--             , Nexus.testSuite
              , TNT.testSuite
              ]