+++ lib/file-parsers/test/TestSuite/GeneratedTests/Internal.hs
@@ -8,17 +8,26 @@ import System.Directory
 
 
 -- |
+-- The directory containing all the test files.
+pathPrefix :: String
+pathPrefix = "lib/file-parsers/test/data-sets/"
+
+
+-- |
 -- Gets all the given files and their contents in the specified directory
 getFileContentsInDirectory :: FilePath -> IO (Map FilePath String)
 getFileContentsInDirectory path = do
-    exists <- doesDirectoryExist path
+    let sep | head path /= '/' && last pathPrefix /= '/' = "/"
+            | otherwise = ""
+    let fullPath = pathPrefix <> sep <> path
+    exists <- doesDirectoryExist fullPath
     if not exists
     then pure mempty
     else do
-      files  <- filter isFile <$> getDirectoryContents path
-      sequence . fromList $ (id &&& readFile) . withPath <$> files
+      files  <- filter isFile <$> getDirectoryContents fullPath
+      sequence . fromList $ (id &&& readFile) . withPath fullPath <$> files
   where
     isFile = not . all (=='.')
-    withPath file = if last path /= '/'
-                    then path <> "/" <> file
-                    else path <> file
+    withPath p file
+      | last p /= '/' = p <> "/" <> file
+      | otherwise     = p        <> file